move types around, move interfaces to common
