+++ common/types/provides.go
@@ -1,4 +1,4 @@
-package common
+package types
 
 import "errors"
