+++ common/types/provides_test.go
@@ -1,4 +1,4 @@
-package common
+package types
 
 import (
 	"testing"
